requirements.txt
numpy>=1.22.0
pandas>=1.5.0
networkx>=2.8.0
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.14.0
umap-learn>=0.5.3
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
spacy>=3.5.0
gradio>=4.21.0
openai>=1.0.0
ipython>=7.0.0
python-dotenv>=0.19.0

# Language models for spaCy
# After installing requirements, run:
# python -m spacy download en_core_web_sm
# python -m spacy download it_core_news_sm (optional) 